Labfiles/uplevel/03-mcp-integration/Python/requirements.txt
python-dotenv>=1.0.0
azure-ai-projects>=1.0.0b4
azure-identity>=1.19.0
pytz>=2024.1
mcp>=1.1.2
